Add IsCompleted flag to scheduled jobs
Introduces an IsCompleted property to JobInfo and ScheduledTask to indicate when a scheduling loop has fully exited (i.e., the job is fully cancelled). Updates the scheduler service to set this flag and include it in job snapshots and hashtable outputs. Also adds an asynchronous CancelAsync method to allow waiting for job cancellation.